Move housing company tests to a single file
